Users edits

+++ app/Http/Resources/OrderMenuResource.php
@@ -15,7 +15,7 @@ class OrderMenuResource extends JsonResource
             'menu_item' => new MenuItemResource($this->whenLoaded('menuItem')),
             'qty_ordered' => $this->qty_ordered,
             'item_total' => $this->qty_ordered * $this->menuItem->price,
-            'formatted_item_total' => number_format($this->qty_ordered * $this->menuItem->price, 2),
+            'formatted_item_total' => number_format((float) ($this->qty_ordered * $this->menuItem->price), 2),
         ];
     }
 }
